Opdrachten gemaakt

+++ src/com/H01/Opdracht6.java
@@ -2,7 +2,6 @@ package com.H01;
 
 public class Opdracht6 {
     public static void main(String[] args) {
-        //werkt niet, nog fiksen
-        System.out.println(4 * (1 - (1/3 + 1/5 - 1/7 + 1/9 - 1/11 + 1/13)));
+        System.out.println(4 * (1 - (double) 1 / 3 + (double) 1 / 5 - (double) 1 / 7 + (double) 1 / 9 - (double) 1 / 11 + (double) 1 / 13));
     }
 }
